refactor: create graph directory
