+++ packages/skott/src/cache/handler.ts
@@ -3,7 +3,8 @@ import path from "node:path";
 import type { FileReader } from "../filesystem/file-reader.js";
 import type { FileWriter } from "../filesystem/file-writer.js";
 import type { Logger } from "../logger.js";
-import type { SkottConfig, SkottNode, SkottStructure } from "../skott.js";
+import type { SkottConfig, SkottStructure } from "../skott.js";
+import type { SkottNode } from "../graph/node.js";
 
 import { createNodeHash, isConfigurationAffected } from "./affected.js";
