+++ packages/skott/src/graph/node.ts
@@ -1,0 +1,9 @@
+import { VertexDefinition } from "digraph-js";
+
+export type SkottNodeBody = {
+  size: number;
+  thirdPartyDependencies: string[];
+  builtinDependencies: string[];
+};
+
+export type SkottNode<T = unknown> = VertexDefinition<SkottNodeBody & T>;
